feat: hide userId on the votes on cards and mergedcards
